\name{tp368-package}
\alias{tp368-package}
\alias{tp368}
\docType{package}
\title{
\packageTitle{tp368}
}
\description{
\packageDescription{tp368}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tp368}
\packageIndices{tp368}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{tp368}

Maintainer: \packageMaintainer{tp368}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
